third_party/WebKit/LayoutTests/fast/compositorworker/compositor-proxy-disconnect.html

<!DOCTYPE html>
<div id="container"></div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

test(function() {
    var first = new CompositorWorker('resources/proxy-echo.js');
    var proxy = new CompositorProxy(document.getElementById('container'), ['opacity']);
    first.postMessage(proxy);

    proxy.disconnect();
    assert_throws("DataCloneError", function() { first.postMessage(proxy); });
}, "This test checks that an element's compositor proxy can't be sent across via post-message after disconnect().");

</script>
